% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_regression.R
\name{assay_regression}
\alias{assay_regression}
\title{Standard Curve Polynomial Regression Fitting}
\usage{
assay_regression(data, concentration, absorbance, stats = FALSE)
}
\value{
A regression equation with protein concentration as y and absorbance as x.
}
\description{
A function that returns the polynomial regression equations for BCA/ Bradford standard curves.
}
\examples{
sample_data <- bca_bradford_bsa_standard_curve
assay_regression(data = sample_data, concentration = Concentration.ug.mL., absorbance = Bradford_Absorbance, stats = TRUE)

}
