% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayAnalysis.R
\name{assayAnalysis}
\alias{assayAnalysis}
\title{Full Analysis of Raw BCA/ Bradford Assay Data}
\usage{
assayAnalysis(
  assay,
  stdcurve,
  stdconc,
  stdabs,
  data,
  concentration = NULL,
  absorbance,
  stats = FALSE,
  printWell = TRUE
)
}
\arguments{
\item{assay}{Name of assay performed as a string: BCA, Bradford, or UV 280}

\item{stdcurve}{List containing  absorbance and concentration values obtained in creation of a standard curve.}

\item{stdconc}{Name of column containing concentration values in stdcurve.}

\item{stdabs}{Name of column containing absorbance.}

\item{data}{Dataframe containing absorbance and concentration values.}

\item{concentration}{Name of column containing concentration values in data.}

\item{absorbance}{Name of column containing absorbance values in data.}

\item{stats}{True or False.}

\item{printWell}{True or False.}
}
\value{
A plot showing the data points of the standard curve, the fitted polynomial regression model of that standard curve, and the concentrations of experimental samples.
}
\description{
A function that returns fitted standard curve regression models and calculated concentrations, given standard curve and experimental absorbance data for BCA and Bradford Assays.
}
\examples{
\dontrun{
}
}
