% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayConc.R
\name{assayConc}
\alias{assayConc}
\title{Concentrations for Unknown Protein Samples}
\usage{
assayConc(data, absorbance, regression, printWell = TRUE)
}
\value{
A data frame with the concentrations of each sample.
}
\description{
A function that returns the concentrations of unknown protein samples, based on absorbance data.
}
\examples{
data(bca_bradford_bsa_standard_curve)
sample_data <- bca_bradford_bsa_standard_curve
bradford_sample <- bradford_sample_data
sample_bradf_regression <- assay_regression(data = sample_data, concentration = Concentration.ug.mL., absorbance = Bradford_Absorbance)
assayConc(data = bradford_sample, absorbance = Absorbance, regression = sample_bradf_regression)

}
