% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayGraph.R
\name{assayGraph}
\alias{assayGraph}
\title{Standard Curve Graphing for UV280}
\usage{
assayGraph(data, absorbance, concentration)
}
\value{
A plot showing the data points of the standard curve, as well as the fitted Beer's Law equation.
}
\description{
A function that returns a graph of the Beer's Law equation for UV280 standard curves.
}
\examples{
\dontrun{
sample_bradf <- assayConc(data = bradford_sample_data, absorbance = Absorbance, regression = sample_bradf_regression)
assayGraph(data = sample_bradf, absorbance = Absorbance, concentration = Concentration)
}

}
