% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_bradford_bsa_standard_curve.R
\name{bca_bradford_bsa_standard_curve}
\alias{bca_bradford_bsa_standard_curve}
\title{Protein Quantification Tests BCA and Bradford}
\format{
Data frame with concentration and absorbance values.
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data used in the creation of BSA standard curves for BCA and Bradford assays.
}
\examples{
bca_bradford_bsa_standard_curve

}
