% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCA_sample_data.R
\docType{data}
\name{BCA_sample_data}
\alias{BCA_sample_data}
\title{Protein Quantification Test BCA_sample_data}
\format{
Data frame with concentration and absorbance values.
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\usage{
BCA_sample_data
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data used in the creation of a BSA standard curve for the BCA protein quantification assay.
}
\examples{
BCA_sample_data

}
\keyword{datasets}
