% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradford_SampleData.R
\name{bradford_SampleData}
\alias{bradford_SampleData}
\title{Sample Bradford Absorbances}
\format{
Data frame with absorbance values.
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data taken from performing a Bradford assay on 7 test protein solutions of unknown concentration.
}
\examples{
bradford_sample_data

}
