% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_tool.R
\name{graphing_tool}
\alias{graphing_tool}
\title{Standard Curve Graphing}
\usage{
graphing_tool(data, concentration, absorbance)
}
\arguments{
\item{data}{Dataframe containing absorbance and concentration values.}

\item{concentration}{Name of column containing concentration values in data.}

\item{absorbance}{Name of column containing absorbance values in data.}
}
\value{
A plot showing the data points of the standard curve, as well as the fitted polynomial regression model.
}
\description{
A function that returns a graph of the polynomial regression fitting for BCA and Bradford standard curves.
}
\examples{
data(bca_bradford_bsa_standard_curve)
sample_data <- bca_bradford_bsa_standard_curve
graphing_tool(data = sample_data, concentration = Concentration.ug.mL., absorbance = Bradford_Absorbance)

}
