% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Protein Quantification Tests BCA Bradford and UV280}
\format{
Data frames with concentration and absorbance values
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data used in the creation of BSA standard curves for three protein quantification methods: BCA assay, Bradford assay, and UV280. It also provides method-specific data of absorbance values across 7 test proteins.
}
\examples{
sample_data <- bca_bradford_bsa_standard_curve
assay_regression(data = sample_data,
                 concentration = Concentration.ug.mL.,
                 absorbance = Bradford_Absorbance,
                 stats = TRUE)

}
