% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve_UV280.R
\docType{data}
\name{standard_curve_UV280}
\alias{standard_curve_UV280}
\alias{UV_bsa_standard_curve}
\title{Protein Quantification Test UV280}
\format{
Data frame with concentration and absorbance values.
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\usage{
UV_bsa_standard_curve
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data used in the creation of a BSA standard curve for the UV280 protein quantification method.
}
\examples{
UV_bsa_standard_curve

}
\keyword{datasets}
