% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UV280_regression.R
\name{UV280_regression}
\alias{UV280_regression}
\title{Standard Curve Linear Regression Fitting}
\usage{
UV280_regression(data, concentration, absorbance, stats = FALSE)
}
\arguments{
\item{data}{Dataframe containing absorbance and concentration values.}

\item{concentration}{Name of column containing concentration values in data.}

\item{absorbance}{Name of column containing absorbance values in data.}

\item{stats}{True or False.}
}
\value{
A regression equation with protein concentration as y and absorbance as x.
}
\description{
A function that returns the linear regression equations for UV 280 standard curves.
}
\examples{
data(UV_bsa_standard_curve)
sample_data <- UV_bsa_standard_curve
UV280_regression(sample_data, Concentration.ug.mL., Absorbance, stats = TRUE)

}
