% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UV_graphing_tool.R
\name{UV_graphing_tool}
\alias{UV_graphing_tool}
\title{Standard Curve Graphing}
\usage{
UV_graphing_tool(data, concentration, absorbance)
}
\arguments{
\item{data}{Dataframe containing absorbance and concentration values.}

\item{concentration}{Name of column containing concentration values in data.}

\item{absorbance}{Name of column containing absorbance values in data.}
}
\value{
A plot showing the data points of the standard curve, as well as the fitted linear regression model.
}
\description{
A function that returns a graph of the linear regression fitting for UV 280 standard curve.
}
\examples{
data(UV_bsa_standard_curve)
sample_data <- UV_bsa_standard_curve
UV_graphing_tool(data = sample_data, concentration = Concentration.ug.mL., absorbance = Absorbance)

}
