% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UV_sample_data.R
\docType{data}
\name{UV_sample_data}
\alias{UV_sample_data}
\title{Protein Quantification Test UV Sample}
\format{
Data frame with concentration and absorbance values.
}
\source{
R. Jiang and A.Y. Cheng (2026), BCH 253-01 Smith College
}
\usage{
UV_sample_data
}
\description{
The data was collected in an introductory Biochemistry lab course at Smith College in the Spring 2026 semester. It includes absorbance data used in the creation of a BSA standard curve for the UV280 protein quantification method.
}
\examples{
UV_sample_data

}
\keyword{datasets}
