% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_regression.R
\name{assay_regression}
\alias{assay_regression}
\title{Standard Curve Polynomial Regression Fitting}
\usage{
assay_regression(data, concentration, absorbance, stats = TRUE)
}
\arguments{
\item{data}{Dataframe containing absorbance and concentration values.}

\item{concentration}{Name of column containing concentration values in data.}

\item{absorbance}{Name of column containing absorbance values in data.}

\item{stats}{True or False.}
}
\value{
Unless stats is set to FALSE, will return a regression equation with protein concentration as y and absorbance as x.
}
\description{
A function that returns the polynomial regression equations for BCA/ Bradford standard curves.
}
\examples{
data(bca_bradford_bsa_standard_curve)
sample_data <- bca_bradford_bsa_standard_curve
assay_regression(sample_data, Concentration.ug.mL., Bradford_Absorbance)

}
