% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayConc.R
\name{assayConc}
\alias{assayConc}
\title{Concentrations for Unknown Protein Samples}
\usage{
assayConc(data, absorbance, regression, printWell = TRUE)
}
\arguments{
\item{data}{*add description}
}
\value{
A data frame with the concentrations of each sample.
}
\description{
A function that returns the concentrations of unknown protein samples, based on absorbance data.
}
\examples{
sample_bradf_regression <- assay_regression(data = bca_bradford_bsa_standard_curve, concentration = Concentration.ug.mL., absorbance = Bradford_Absorbance)
assayConc(data = bradford_sample_data, absorbance = Absorbance, regression = sample_bradf_regression)

}
