% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayConc.R
\name{assayConc}
\alias{assayConc}
\title{Concentrations for Unknown Protein Samples}
\usage{
assayConc(data, absorbance, regression, printWell = TRUE)
}
\arguments{
\item{data}{List containing absorbance values.}

\item{absorbance}{Name of column containing absorbance values in data.}

\item{regression}{List of regression modeling coefficients.}

\item{printWell}{True or False.}
}
\value{
A data frame with the concentrations of each sample.
}
\description{
A function that returns the concentrations of unknown protein samples, based on absorbance data, as a data frame.
}
\examples{
sample_bradf_regression <- assay_regression(bca_bradford_bsa_standard_curve, 
Concentration.ug.mL., 
Bradford_Absorbance)
assayConc(bradford_sample_data, Absorbance, sample_bradf_regression)

}
